Fix lock screen running CES2016 DEMO without Navigation binary
When CES2016 DEMO lauched by 'start_CES2016_ivi-shell.sh'
which is for environment without Navigation binary,
the operation of touch/click on icon of navigation will
lock the whole screen and it can't be recovered.

This patch solve this issue by correcting control of
switching navigation to show or hide.

Change-Id: I8ed07e06e5af8da40d8a2386e1a14f2465ce8bf0
Signed-off-by: Tadao Tanikawa <tanikawa.tadao@jp.panasonic.com>